+++ src/bus/index.js
@@ -81,13 +81,19 @@ class Bus {
    * If a connection already exists for that organization id, the connection is returned, otherwise a new connection is created and returned.
    *
    * @param {string} organizationId UUID corresponding with an organization
+   * @param {function} onClose optional callback to be executed when the connection is closed
+   * @param {function} onError optional callback to be executed when the connection encounters an error
    *
    * @returns {Promise}
    * @fulfill {WebSocketConnection}
    * @reject {errorEvent} The error event
    *
    * @example
-   * contxtSdk.bus.connect('4f0e51c6-728b-4892-9863-6d002e61204d')
+   * contxtSdk.bus.connect('4f0e51c6-728b-4892-9863-6d002e61204d', (orgId, evt) => {
+   *   console.log(`connection closed: ${evt}`);
+   * }, (orgId, evt) => {
+   *   console.log(`connection error: ${evt}`);
+   * })
    *   .then((webSocket) => {
    *     console.log(webSocket);
    *   })
@@ -95,7 +101,7 @@ class Bus {
    *     console.log(errorEvent);
    *   });
    */
-  connect(organizationId) {
+  connect(organizationId, onClose, onError) {
     return new Promise((resolve, reject) => {
       if (this._webSockets[organizationId]) {
         return resolve(this._webSockets[organizationId]);
@@ -124,13 +130,31 @@ class Bus {
             resolve(this._webSockets[organizationId]);
           };
 
-          ws.onclose = (event) => {
+          ws.addEventListener("close", (event) => {
             this._webSockets[organizationId] = null;
-          };
+            if (onClose) {
+              try {
+                onClose(organizationId, event);
+              } catch (ex) {
+                console.log('Message Bus Error calling onClose callback: ', ex)
+              }
+            }
+          });
 
-          ws.onerror = (errorEvent) => {
-            reject(errorEvent);
-          };
+          ws.addEventListener("error", (errorEvent) => {
+            const connected = this._webSockets[organizationId] != null;
+            this._webSockets[organizationId] = null;
+            if (!connected) {
+              reject(errorEvent)
+            }
+            if (onError) {
+              try {
+                onError(organizationId, errorEvent);
+              } catch (ex) {
+                console.log('Message Bus Error calling onError callback: ', ex)
+              }
+            }
+          });
         })
         .catch((err) => {
           reject(err);