+++ src/bus/webSocketConnection.js
@@ -38,7 +38,7 @@ class WebSocketConnection {
     this._autoAck = autoAcknowledge;
 
     if (this._webSocket) {
-      this._webSocket.onerror = this._onError;
+      this._webSocket.addEventListener("error", this._onError);
       this._webSocket.onmessage = this._onMessage;
     }
   }